+++ scripts/test.sh
@@ -452,6 +452,7 @@ compile_and_expect containers 42
 compile_and_expect cpp_template_interop 42; compile_and_expect cpp_move_unique_ptr 42; compile_and_expect cpp_filesystem_path 42; compile_and_expect cpp_chrono_timer 42
 compile_and_expect cpp_template_member 42
 compile_and_expect cpp_operator_overload 42
+compile_and_expect dudu_operator_overload 42
 compile_and_expect cpp_overloaded_constructor 42
 compile_and_expect cpp_digit_underscore_name 42
 compile_and_expect std_vector_map_string 42